requirements.txt
streamlit
langchain
langgraph
langchain-community
langchain-core
langchain-groq
langchain-huggingface
sentence-transformers
faiss-cpu
python-dotenv
requests
pypdf
duckduckgo-search
